Move "make-info" key to a separate file
Elm manages the "elm[-package].json" file automatically, so extra keys
added to the file are likely to get lost when Elm makes an update to
the file.  Instead of trying to maintain the special "make-info" key
in the Elm package file, this establishes a separate hidden file named
".elm-light.json" to store any custom keys for the `elm-light` plugin.

This represents one possible way to address the issue reported in
rundis/elm-light#18.